Try to have a better query starting point
This commit introduces a new `QueryType` trait that allows you to call a
static `::query()` method on types that implement `HasTable` and
`Selectable<DB>`. This is usually the case if you derive `Selectable` +
`Identifiable`. 

The main motivation for this is two fold:
* It's shorter and removes some duplication as the type usually already
  needs to know the table for code generation reasons
* It removes potential failure points. Instead of requiring users to
  call an additional method (`.select(Type::as_select())`) we could give
them a shorter solution that looks easier and does not require calling
additional functions to get the behaviour they expect.

This is just a minimal implementation. The trait is designed in such way
that we could in the future provide an additional derive that allow more
complicated from clauses than "Just that table", like joining etc. This
would require some additional design work and we would need to decide
how exactly that should work. We could have yet another partial derive
you need to put on your struct or we could have some sort of "super"
derive that essentially does all: Deriving `Selectable`, `Identifiable`
(at least the `HasTable` part) and the new trait. The later approach
would have the advantage of simplifying what users need to usually do.
We could still leave the other derives in some soft deprecated state
around, as there are still special cases that could be better solved
with them.