Setup cargo-semver-checks
This commit sets up a xtask for cargo-semver-checks and integrates it in
our CI. It also fixes 2 instances of reported breakage. With that there
are still 3 more instances reported. I believe two of them are false
positives (The `Downcast` super trait and the `IntoArrayExpression`
trait message). The third one is a "real" breakage, but I would
nevertheless say that this will likely not affect anyone as I do not see
any reasonable way to to do something with a trait object there at all.

I'm still not sure what's the correct way to supress such errors only at
specific locations.