Refactor sqlite version check
This change refactors the way we check for certain SQLite versions in
our doc tests. Instead of manually copying the same code over and over
again into different doc tests, I've moved the relevant code into the
`doc_setup.rs` file and provided a helper function + macro for this
check. This makes it much easier to use the same change for different
versions in different locations. 

This PR also copies the version used in the doc tests as minimal
supported versions for that function into the documentation. I did not
check for each version if it's correct. If it turns out to be wrong we
can always fix it later.