Fix typos
This change fixes new issues found by typos

There is one real issue found by typos. I've desided to ignore `tpe` as
that's commonly used in place of `type` in our codebase. We won't change
it as it's hard to use `type` in rust code as it's an reserved
identifier.