Add nlnet and prototype fund logos
This change adds entities that currently support diesel's development
with significant sponsorships/funds as notable sponsors on the web page.
For ProtoType fund and the Federal Ministry of Research that's a
requirement for their funding. For the other entities it's just fair to
have them there as well as they contribute similar amount of funding.