Replace trybuild with ui_test
ui_test is one of the crates used by rustc for compile_tests. I choose
to replace trybuild with ui_test as the former does not perform const
evaluation for the tests, which results in certain patterns not beeing
testable. It also does not seem to be planned to fix that in trybuild. 
See for details https://github.com/dtolnay/trybuild/issues/225

Another thing that I see as advantage of ui_test is that it requires us
to annotate which lines emitt errors, which makes it harder to
accidently change something from not compiling to compiling by accident.

This is a large diff due to that the output of all tests change
slightly. I tried my best to adjust the output via filters us much as
possible to match the trybuild output, but there are still some
differences. If it turns out to be problematic in the future we can
continue to tweak the output.

Finally I also added some docs for the compile tests to the CONTRIBUTING
manual.