Deny using backend specific SQL functions on other backends
This commit denies using backend specific SQL functions on other
backends. Technically that is a breaking change as some of these
function were usable on diesel 2.2.x or earlier. Practially speaking I
would consider that as a bug fix as it just prevents additional queries
to not compile that would otherwise fail at runtime.