Fix .try_into() regression
While cleaning up how we call methods in our proc-macros I broke a
pattern that seemed to be used. This commit restores using a
`.try_into()` call instead of a fully qualified trait call for
converting fields as part of `#[derive(Queryable)]` this unblocks using
a method on the type itself. 
This commit also adds comments, tests + documentations for this use-case
to make sure we don't break it again.