Include expanded code for derives in our documentation
This change includes the expanded code for all our derives in our API
documentation. This makes it hopefully a bit easier for our users to
understand what the macros actually do and also to estimate how much
code they generate.

Technically we just take the snapshots from the tests #4701 and use a
build script to generate the relevant documentation snippets. The output
of the build script is then included into the documentation. It's a bit
unfortune that we cannot check in the build script if it's run by
rustdoc vs some other compilation feature so I decided to gate this on
the `docsrs` cfg flag to not run that step in "normal" builds.

This change also enables the `--generate-link-to-definition` option for
rustdoc.