Fix various issues in the mysql backend
This change improves some of the unsafe code in the mysql backend. These
places were reported as potential sub-optimal during a security review
funded by NGI Core Zero as part of the project for adding support for window
function to diesel.

None of the changes fix a critical exploitable problem as far as I can
tell, it's nevertheless a worth improvement.