Fix yet another mysql_time issue
Turns out that `libmariadb` ships a different definition of `MYSQL_TIME`
than `libmysqlclient` (and diesel). The first variant misses the
`time_zone_displacement` field, which results in the size check failing
as the provided buffer is too small. Now on the one hand that's good
because it would have been a out of bounds read otherwise, on the other
hand it's bad as we need to have yet another workaround for yet another
incompatiblity. This issue also existed in diesel versions older than
2.3, although it did not cause any error there. Strictly speaking we
performed an out of bound read there, although fortunally only of memory
we initialized to be zero before as the underlying buffer used the size
of the diesel `MysqlTime` type to set that many bytes to 0. So that's
still UB, but not a security problem.

The fix for this is essentially to make sure we only copy as many bytes
as there are there. We set the other bytes of the struct to zero by
initializing everything explizitly to zero first. I've also choosen to
introduce another check for the `neg` field of the `MysqlTime` struct as
that's the only field that has requirements on the values there. For any
other field any bit value is valid. For `neg` only 0 and 1 are valid
values due to the `bool` field value. Any other value would have been
UB. By checking we prevent ever reaching that point.

Fixes #4765