Fix cargo deny
This adds a deny.toml file that configures which licenses are allowed.

This also enables a few more deny settings to deny various stuff during
builds. I would like to enable more of the settings but some of the
dependencies fail them. We likely need to fix these dependencies first.