Fix generated function links
The links on the return type helper back to the function generated by `#[declare_sql_function]` were broken. This commit fixes these links.

See the links in our documentation:
https://docs.diesel.rs/main/diesel/sqlite/expression/dsl/index.html#types
for broken examples. For example the `json()` link at the `json` type
alias is expected to link back to the `json` function defined a bit
above.