Refactor trait setup to put `retrieve_as_text` json operator into a shared trait
This commit refactors the json operator trait setup to move the method
that's the same between postgres and sqlite into a shared trait and also
ensures that the types are shared between this function.

Additionally it renames the `retrieve_as_object` method for sqlite to
fix the name collision with the postgres method.

Finally this commit also fixes that the `retrieve_as_object` method for
sqlite always returns json not jsonb.