Remove a misleading documentation comment for concurrent usage of `run_pending_migrations`
This commit removes a comment that states running
`run_pending_migrations` concurrently was actually safe. This is not the
case as we e.g don't lock the migrations table to prevent other
processes running migrations itself. In the long run we likely want to
implment a proper fix for this, for now it's better to just correct the
documentation.

Fix 4862