Extend diesel-infer-query to cover more complex queries
This PR extends diesel-infer-query to cover more complex SQL
expressions. This includes:

- `IN` (including subqueries)
- `NOT IN` (including subqueries)
- Operator nesting
- Plain subqueries
- Common table expressions
- Nested common table expressions
- Casting
- Using subqueries as query source
- Set Queries like `UNION`, `INTERSECT` and `EXCEPT`

This also refactors some part of the public API of `diesel-infer-query`
to use an explict three value enum to indicate if a expression in
nullable or not (or unknown).

If during review nothing critical comes up I think that covers the
remaining parts of #43

Fixes #43