Add no_std support to diesel
This commit adds support for using diesel with the SQLite backend
 in a no-std (core + alloc) environment. 

This mainly means fixing a lot of imports to use the relevant items from
core or alloc instead. Notable exceptions are:

* There is no hashmap in alloc, so we need to pull in hashbrown for this
* There is no panicing/abort support in core so we need our custom shims
  for these.
* Global instrumentation setting is not supported yet (cfged away) as it
  requires lock which are not supported on embedded targets.

This commit also introduces a new `std` default feature for enabling `std` lib
support. If disabled the users need to manually enable the `hashbrown`
feature to pull in hashbrown for hashmap support. Cargo doesn't support
pulling in dependencies for disabled features.

Finally this commit also adds a `example/sqlite/embedded` example
project that demonstrates how to use this on a esp32-c6.