Another mariadb fix
This commit fixes another issue with libmariadb. In this case again the
libmariadb documentation/behaviour is inconsistent. They stated that
with the last version they won't set the lenght field for statically
sized types. Turns out that even that is not correct, as it is set for
the datetime type. Now this on it's own wouldn't be an issue, but
libmariadb has a slightly different definition of `MYSQL_TIME` that
libmysqlclient. It misses a field in the end. While this is normally not
a problem as that field is just zero initialized by diesel, it becomes
problematic if libmariadb returns the size of their definition, which
results in the assert mentioned in the issue to fail. So it's yet
another needless incompatiblity on their side we need to workaround.

Fixes #4889